fix: better bar chart colors
