+++ packages/common-utils/src/clickhouse.ts
@@ -494,158 +494,6 @@ export function inferTimestampColumn(
   return filterColumnMetaByType(meta, [JSDataType.Date])?.[0];
 }
 
-function inferValueColumns(meta: Array<{ name: string; type: string }>) {
-  return filterColumnMetaByType(meta, [JSDataType.Number]);
-}
-
-function inferGroupColumns(meta: Array<{ name: string; type: string }>) {
-  return filterColumnMetaByType(meta, [
-    JSDataType.String,
-    JSDataType.Map,
-    JSDataType.Array,
-  ]);
-}
-
-// TODO: Move to ChartUtils
-// Input: { ts, value1, value2, groupBy1, groupBy2 },
-// Output: { ts, [value1Name, groupBy1, groupBy2]: value1, [...]: value2 }
-export function formatResponseForTimeChart({
-  res,
-  dateRange,
-  granularity,
-  generateEmptyBuckets = true,
-}: {
-  dateRange: [Date, Date];
-  granularity?: SQLInterval;
-  res: ResponseJSON<Record<string, any>>;
-  generateEmptyBuckets?: boolean;
-}) {
-  const meta = res.meta;
-  const data = res.data;
-
-  if (meta == null) {
-    throw new Error('No meta data found in response');
-  }
-
-  const timestampColumn = inferTimestampColumn(meta);
-  const valueColumns = inferValueColumns(meta) ?? [];
-  const groupColumns = inferGroupColumns(meta) ?? [];
-
-  if (timestampColumn == null) {
-    throw new Error(
-      `No timestamp column found with meta: ${JSON.stringify(meta)}`,
-    );
-  }
-
-  // Timestamp -> { tsCol, line1, line2, ...}
-  const tsBucketMap: Map<number, Record<string, any>> = new Map();
-  const lineDataMap: {
-    [keyName: string]: {
-      dataKey: string;
-      displayName: string;
-      maxValue: number;
-      minValue: number;
-      color: string | undefined;
-    };
-  } = {};
-
-  for (const row of data) {
-    const date = new Date(row[timestampColumn.name]);
-    const ts = date.getTime() / 1000;
-
-    for (const valueColumn of valueColumns) {
-      const tsBucket = tsBucketMap.get(ts) ?? {};
-
-      const keyName = [
-        valueColumn.name,
-        ...groupColumns.map(g => row[g.name]),
-      ].join(' · ');
-
-      // UInt64 are returned as strings, we'll convert to number
-      // and accept a bit of floating point error
-      const rawValue = row[valueColumn.name];
-      const value =
-        typeof rawValue === 'number' ? rawValue : Number.parseFloat(rawValue);
-
-      tsBucketMap.set(ts, {
-        ...tsBucket,
-        [timestampColumn.name]: ts,
-        [keyName]: value,
-      });
-
-      // TODO: Set name and color correctly
-      lineDataMap[keyName] = {
-        dataKey: keyName,
-        displayName: keyName,
-        color: undefined,
-        maxValue: Math.max(
-          lineDataMap[keyName]?.maxValue ?? Number.NEGATIVE_INFINITY,
-          value,
-        ),
-        minValue: Math.min(
-          lineDataMap[keyName]?.minValue ?? Number.POSITIVE_INFINITY,
-          value,
-        ),
-      };
-    }
-  }
-
-  // TODO: Custom sort and truncate top N lines
-  const sortedLineDataMap = Object.values(lineDataMap).sort((a, b) => {
-    return a.maxValue - b.maxValue;
-  });
-
-  if (generateEmptyBuckets && granularity != null) {
-    // Zero fill TODO: Make this an option
-    const generatedTsBuckets = timeBucketByGranularity(
-      dateRange[0],
-      dateRange[1],
-      granularity,
-    );
-
-    generatedTsBuckets.forEach(date => {
-      const ts = date.getTime() / 1000;
-      const tsBucket = tsBucketMap.get(ts);
-
-      if (tsBucket == null) {
-        const tsBucket: Record<string, any> = {
-          [timestampColumn.name]: ts,
-        };
-
-        for (const line of sortedLineDataMap) {
-          tsBucket[line.dataKey] = 0;
-        }
-
-        tsBucketMap.set(ts, tsBucket);
-      } else {
-        for (const line of sortedLineDataMap) {
-          if (tsBucket[line.dataKey] == null) {
-            tsBucket[line.dataKey] = 0;
-          }
-        }
-        tsBucketMap.set(ts, tsBucket);
-      }
-    });
-  }
-
-  // Sort results again by timestamp
-  const graphResults: {
-    [key: string]: number | undefined;
-  }[] = Array.from(tsBucketMap.values()).sort(
-    (a, b) => a[timestampColumn.name] - b[timestampColumn.name],
-  );
-
-  // TODO: Return line color and names
-  return {
-    // dateRange: [minDate, maxDate],
-    graphResults,
-    timestampColumn,
-    groupKeys: sortedLineDataMap.map(l => l.dataKey),
-    lineNames: sortedLineDataMap.map(l => l.displayName),
-    lineColors: sortedLineDataMap.map(l => l.color),
-  };
-}
-
 export type ColumnMeta = {
   codec_expression: string;
   comment: string;
